@@ -17,9 +17,9 @@
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
-    "react": "^17.0.0",
+    "react": "^18.0.0",
     "react-activity": "^2.1.3",
-    "react-dom": "^17.0.0",
+    "react-dom": "^18.0.0",
     "react-dropzone": "^11.4.2",
     "react-query": "^3.32.1",
     "react-redux": "^7.2.6",
@@ -30,8 +30,8 @@
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.0",
     "@types/lodash": "^4.14.176",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
     "@typescript-eslint/parser": "^5.13.0",
     "@vitejs/plugin-react": "1.0.5",
